Topic/beta build (#51)
* temp full import of users

* temp full import of users

* remove temp import

* guild track

* guild track

* tweak live now

* remove duplicate activites from before/after

* trivia tracking

* added trivia initiator

* remove duplicate activites from before/after

* add guild info to game keys and minecraft whitelist

* added logging for live now post

* added logging for live now post

* change when the roles are updated

* move message exit if not right emoji

* use utils.get_user_displayname

* log method source

* some code cleanup

* code cleanup

* code cleanup

* do a cleanup at the end

* cleanup code to use function vs duplicated code

* exit if in DM for food photo

* better settings check for streamteam cog

* archive taco tuesday message. add user to taco tuesday role

* fixed the move message. auto add taco tuesday when using the 'new' command

* only add reaction fields if there are reactions

* fix CHOICE_EMOJIS_DEFAULTS

* add info to log values

* fix cleanup

* format logging

* more logging cleaning

* more logging cleaning

* fix live now

* code cleanup

* rename sendEmbed to send_embed

* use display name for author set embeds

* more logging consistency updates

* update cogs with consistent logging names

* more logging consistency updates

* some method typing

* update python version

* python 3.11

* rename updateEmbed to update_embed

* fix call to get tacos settings

* update taco types

* add recieve tacos type

* added user join migration step

* remove the migrate user join

* remove the migrate user join

* use default avatar if doesnt exist

* taco_tuesday migrate start

* update message

* added taco tuesday import

* fix import call optional arg

* check if user is bot

* only admin can import / archive

* check if user is bot

* check admin early

* fix bot

* fix get role

* fix bot

* describe the taco tuesday commands for docs. update the reaction events for wdyctw, tech thurs, mental monday to only trigger for admin

* update to use get_displayname when displaying user with discriminator

* refactor adding twitch user info to remove the twitch_user_id as it was never populated.

* remove 'asking' for twitch name when joining streamer role

* added permissions helper

* Fix mental monday missing _method variable

* some code cleanup

* kick 'too new' accounts

* added system actions tracking

* remove action

* update the app manifest

* track twitch discord link actions

* added some system event tracking

* save action name

* update avatar for user update

* await cog add

* get count of reactions

* get count of reactions

* check data for system action

* upgrade discordpy

* track user status

* get member if it is a user object

* get reaction count

* catch failure of get_fetch_member

* remove warning log for user not found

* remove update member on message

* delete user messages from system channel when kicking them out on 'new account check'

* log when cannot delete kick messages

* fix call to load permissions object

* permissions helper

* use permissions helper is_admin function

* do not store 'None' for id's

* await coroutine

* tweak the user join whitelist

* added command sync

* track invites as system action

* track invite code in the system action

* rename food-photo

* handle photo posts

* migrate taco log type for photo phost

* Added voice chat tacos

* added intro import

* fix import

* handle timeout of gamekey

* set temp 60 second timeout to test game key

* reset temp timeout

* send_embed refactor out of 'discordhelper'

* fix on-message in intro to exit if we are not in a guild

* added live now logging

* added logging for re-attaching game-key

* try to import the view.

* fix live now tracking

* fix add view

* cleanup live after there is no more streaming activities for a user

* automatically add a new game key offer when the discord is available.

* remove auto game key

* remove the second defer

* log interaction handler

* log interaction handler

* added logging

* remove listener

* fix the claiming of games

* give the tacos to the person that donated the game

* handle interaction if the original listener fails

* handle interaction when 'disconnected' so it can still be processed.

* move game cost over to the game key vs being set for all games

+++ bot/cogs/live_now.py
@@ -26,6 +26,7 @@ from .lib import settings
 from .lib import mongo
 from .lib import tacotypes
 from .lib.system_actions import SystemActions
+from .lib.messaging import Messaging
 
 class LiveNow(commands.Cog):
     def __init__(self, bot):
@@ -35,6 +36,7 @@ class LiveNow(commands.Cog):
         self.bot = bot
         self.settings = settings.Settings()
         self.discord_helper = discordhelper.DiscordHelper(bot)
+        self.messaging = Messaging(bot)
         self.SETTINGS_SECTION = "live_now"
         self.db = mongo.MongoDatabase()
         log_level = loglevel.LogLevel[self.settings.log_level.upper()]
@@ -77,23 +79,25 @@ class LiveNow(commands.Cog):
 
             for bsa in before_streaming_activities_temp:
                 # if item is not in the list, add it
-                if len([a for a in before_streaming_activities if a.platform == bsa.platform]) == 0:
+                if len([b for b in before_streaming_activities if b.platform.lower() == bsa.platform.lower()]) == 0:
+                    self.log.debug(guild_id, f"{self._module}.{_method}", f"Adding {bsa.platform} to before_streaming_activities for {before.display_name}")
                     before_streaming_activities.append(bsa)
 
             for asa in after_streaming_activities_temp:
-                if len([a for a in after_streaming_activities if a.platform == asa.platform]) == 0:
+                if len([a for a in after_streaming_activities if a.platform.lower() == asa.platform.lower()]) == 0:
+                    self.log.debug(guild_id, f"{self._module}.{_method}", f"Adding {asa.platform} to after_streaming_activities for {after.display_name}")
                     after_streaming_activities.append(asa)
 
             # WENT LIVE
             for asa in after_streaming_activities:
+                await asyncio.sleep(1)
+
                 tracked = self.db.get_tracked_live(guild_id, after.id, asa.platform)
                 is_tracked = tracked != None and tracked.count() > 0
                 # if it is already tracked, then we don't need to do anything
                 if is_tracked:
                     self.log.debug(guild_id, f"{self._module}.{_method}", f"{after.display_name} is already tracked for {asa.platform}")
-
                     await self.add_live_roles(after, cog_settings)
-
                     continue
 
                 self.log.info(guild_id, f"{self._module}.{_method}", f"{after.display_name} started streaming on {asa.platform}")
@@ -114,7 +118,6 @@ class LiveNow(commands.Cog):
                 )
 
                 await self.add_live_roles(after, cog_settings)
-
 
                 twitch_name: typing.Union[str, None] = None
                 if asa.platform.lower() == "twitch":
@@ -144,54 +147,54 @@ class LiveNow(commands.Cog):
 
 
             # ENDED STREAM
-            for bsa in before_streaming_activities:
-                # sleep for a bit to make sure the live role is removed before we clean up the db
-                await asyncio.sleep(1)
+            # HMMMM. This is not triggering. Need to figure out why
+            # for bsa in before_streaming_activities:
+            #     # sleep for a bit to make sure the live role is removed before we clean up the db
+            #     await asyncio.sleep(1)
 
-                # check if bsa is in after_streaming_activities
-                found_bsa = len([a for a in after_streaming_activities if a.url == bsa.url and a.platform == bsa.platform]) > 0
-                if found_bsa:
-                    # this activity exists in both lists, so it is still live
-                    continue
+            #     # check if bsa is in after_streaming_activities
+            #     found_bsa = len([a for a in after_streaming_activities if a.url == bsa.url and a.platform == bsa.platform]) > 0
+            #     if found_bsa:
+            #         self.log.debug(guild_id, f"{self._module}.{_method}", f"{after.display_name} is still live on {bsa.platform}")
+            #         # this activity exists in both lists, so it is still live
+            #         continue
 
-                tracked = self.db.get_tracked_live(guild_id, before.id, bsa.platform)
-                is_tracked = tracked != None and tracked.count() > 0
-                # if it is not tracked, then we don't need to do anything
-                if not is_tracked:
-                    self.log.debug(guild_id, f"{self._module}.{_method}", f"{after.display_name} is not tracked for {bsa.platform}")
-                    await self.remove_live_roles(before, cog_settings)
-                    continue
+            #     tracked = self.db.get_tracked_live(guildId=guild_id, userId=before.id, platform=bsa.platform)
+            #     is_tracked = tracked != None and tracked.count() > 0
+            #     # if it is not tracked, then we don't need to do anything
+            #     if not is_tracked or not tracked:
+            #         self.log.debug(guild_id, f"{self._module}.{_method}", f"{after.display_name} is not tracked for {bsa.platform}")
+            #         await self.remove_live_roles(before, cog_settings)
+            #         continue
 
-                # if we get here, then we need to untrack the user live activity
-                self.log.info(guild_id, f"{self._module}.{_method}", f"{before.display_name} stopped streaming on {bsa.platform}")
-                # track the END activity
-                self.db.track_live_activity(guild_id, before.id, False, bsa.platform, url=bsa.url)
+            #     # if we get here, then we need to untrack the user live activity
+            #     self.log.info(guild_id, f"{self._module}.{_method}", f"{before.display_name} stopped streaming on {bsa.platform}")
+            #     # track the END activity
+            #     self.db.track_live_activity(guild_id, before.id, False, bsa.platform, url=bsa.url)
 
-                logging_channel_id = cog_settings.get("logging_channel", None)
-                if logging_channel_id:
-                    logging_channel = await self.discord_helper.get_or_fetch_channel(int(logging_channel_id))
-                    if logging_channel:
-                        for tracked_item in tracked:
-                            message_id = tracked_item.get("message_id", None)
-                            if message_id:
-                                try:
-                                    message = await logging_channel.fetch_message(int(message_id))
-                                    if message:
-                                        self.log.debug(guild_id, f"{self._module}.{_method}", f"Deleting LIVE 🔴 message ({message_id}) from channel {logging_channel}")
-                                        await message.delete()
-                                except discord.errors.NotFound:
-                                    self.log.warn(guild_id, f"{self._module}.{_method}", f"Message {message_id} not found in channel {logging_channel}")
+            #     logging_channel_id = cog_settings.get("logging_channel", None)
+            #     if logging_channel_id:
+            #         logging_channel = await self.discord_helper.get_or_fetch_channel(int(logging_channel_id))
+            #         if logging_channel:
+            #             for tracked_item in tracked:
+            #                 message_id = tracked_item.get("message_id", None)
+            #                 if message_id:
+            #                     try:
+            #                         message = await logging_channel.fetch_message(int(message_id))
+            #                         if message:
+            #                             self.log.debug(guild_id, f"{self._module}.{_method}", f"Deleting LIVE 🔴 message ({message_id}) from channel {logging_channel}")
+            #                             await message.delete()
+            #                     except discord.errors.NotFound:
+            #                         self.log.warn(guild_id, f"{self._module}.{_method}", f"Message {message_id} not found in channel {logging_channel}")
 
-                # remove all tracked items for this live platform (should only be one)
-                self.db.untrack_live(guild_id, before.id, bsa.platform)
+            #     # remove all tracked items for this live platform (should only be one)
+            #     self.db.untrack_live(guild_id, before.id, bsa.platform)
 
-                await self.remove_live_roles(before, cog_settings)
+            #     await self.remove_live_roles(before, cog_settings)
 
-            # check if the user is in any of the "add" roles, but has no streaming activity
-            # remove them from those roles
             if len(before_streaming_activities) == 0 and len(after_streaming_activities) == 0:
                 try:
-                    await self.remove_live_roles(before, cog_settings)
+                    await self.remove_live_roles(after, cog_settings)
                     await self.clean_up_live(guild_id, after.id)
                 except Exception as e:
                     self.log.error(guild_id, f"{self._module}.{_method}", str(e), traceback.format_exc())
@@ -223,17 +226,29 @@ class LiveNow(commands.Cog):
     def handle_other_live(self, user: discord.Member, activities: typing.List[discord.Streaming]) -> typing.Union[str, None]:
         _method = inspect.stack()[0][3]
         guild_id = user.guild.id
-        if len(activities) > 1:
+        # get non-youtube activities and non-twitch activities
+        other_activities = [a for a in activities if a.platform is not None and a.platform.lower() != "youtube" and a.platform.lower() != "twitch"]
+
+        if len(other_activities) > 1:
             self.log.warn(guild_id, f"{self._module}.{_method}", f"{user.display_name} has more than one streaming activity")
-        activity = activities[0]
-        self.log.info(guild_id, f"{self._module}.{_method}", f"{user.display_name} started streaming on an {activity.platform} platform")
+
+        for a in other_activities:
+            self.log.info(guild_id, f"{self._module}.{_method}", f"{user.display_name} started streaming on an {a.platform} platform")
         return None
 
     def handle_youtube_live(self, user: discord.Member, activities: typing.List[discord.Streaming]) -> typing.Union[str, None]:
         _method = inspect.stack()[0][3]
         guild_id = user.guild.id
-        if len(activities) > 1:
+        # get only the youtube activity
+        youtube_activities = [a for a in activities if a.platform is not None and a.platform.lower() == "youtube"]
+        if len(youtube_activities) > 1:
             self.log.warn(guild_id, f"{self._module}.{_method}", f"{user.display_name} has more than one streaming activity")
+
+        if len(youtube_activities) == 0:
+            self.log.warn(guild_id, f"{self._module}.{_method}", f"{user.display_name} has no youtube streaming activity")
+            return None
+
+        activity = youtube_activities[0]
 
         self.log.info(guild_id, f"{self._module}.{_method}", f"{user.display_name} started streaming on youtube")
         return None
@@ -242,28 +257,23 @@ class LiveNow(commands.Cog):
         _method = inspect.stack()[0][3]
         guild_id = user.guild.id
 
+        # get only the twitch activity
+        twitch_activities = [a for a in activities if a.platform is not None and a.platform.lower() == "twitch"]
+
         twitch_name: str = None
         twitch_info = self.db.get_user_twitch_info(user.id)
-        if len(activities) > 1:
-            self.log.warn(guild_id, "live_now.handle_other_live", f"{user.display_name} has more than one streaming activity")
+        if len(twitch_activities) > 1:
+            self.log.warn(guild_id, f"{self._module}.{_method}", f"{user.display_name} has more than one streaming activity")
         # Only add the twitch name if we don't have it already
         # and only if we can get the twitch name from the url or activity
         # or if we have a different twitch name in the database
-        for activity in activities:
+        for activity in twitch_activities:
             if activity.twitch_name:
                 twitch_name = activity.twitch_name.lower()
                 break
             elif activity.url and "twitch.tv/" in activity.url.lower():
                 twitch_name = activity.url.lower().split("twitch.tv/")[1]
                 break
-            if twitch_name:
-                # track the users twitch name
-                self.db.set_user_twitch_info(user.id, twitch_name.lower())
-                self.db.track_system_action(
-                    guild_id=guild_id,
-                    action=SystemActions.LINK_TWITCH_TO_DISCORD,
-                    data={"user_id": str(user.id), "twitch_name": twitch_name.lower()},
-                )
         if twitch_info:
             twitch_info_name = twitch_info.get("twitch_name", None)
         else:
@@ -321,10 +331,8 @@ class LiveNow(commands.Cog):
                         twitch_name = image_url.replace("twitch:", "")
                         profile_icon = self.get_user_profile_image(twitch_name)
 
-                    # self.log.debug(guild_id, "live_now.log_live_post", f"Found large image {image_url}")
             user_display_name = utils.get_user_display_name(user)
-            self.log.debug(guild_id, f"{self._module}.{_method}", f"Logging live post for {user_display_name} in {logging_channel.name}")
-            message = await self.discord_helper.send_embed(logging_channel,
+            message = await self.messaging.send_embed(logging_channel,
                 f"🔴 {user_display_name}", description,
                 fields, thumbnail=profile_icon,
                 author=user, color=0x6a0dad)
@@ -343,24 +351,39 @@ class LiveNow(commands.Cog):
         if all_tracked_for_user is None or all_tracked_for_user.count() == 0:
             return
 
+        user = await self.discord_helper.get_or_fetch_member(guildId=guild_id, userId=user_id)
+        if user is None:
+            self.log.debug(guild_id, f"{self._module}.{_method}", f"Could not find user {user_id}")
+            return
+
+        self.log.debug(guild_id, f"{self._module}.{_method}", f"Cleaning up {all_tracked_for_user.count()} tracked live items for user {user_id}")
         cog_settings = self.get_cog_settings(guild_id)
 
         logging_channel_id = cog_settings.get("logging_channel", None)
+        logging_channel = None
         if logging_channel_id:
             logging_channel = await self.discord_helper.get_or_fetch_channel(int(logging_channel_id))
-            for tracked in all_tracked_for_user:
-                if logging_channel:
-                    message_id = tracked.get("message_id", None)
-                    if message_id:
-                        try:
-                            message = await logging_channel.fetch_message(int(message_id))
-                            if message:
-                                await message.delete()
-                        except discord.errors.NotFound:
-                            self.log.debug(guild_id, f"{self._module}.{_method}", f"Message {message_id} not found in channel {logging_channel}")
 
-                # remove all tracked items for this live platform (should only be one)
-                self.db.untrack_live(guild_id, tracked.get('user_id'), tracked.get('platform'))
+
+        for tracked in all_tracked_for_user:
+            platform = tracked.get("platform", "UNKNOWN")
+            url = tracked.get("url", None)
+            self.log.info(guild_id, f"{self._module}.{_method}", f"{utils.get_user_display_name(user)} stopped streaming on {platform}")
+            self.db.track_live_activity(guild_id, user.id, False, platform, url=url)
+
+            if logging_channel:
+                message_id = tracked.get("message_id", None)
+                if message_id is not None and message_id != "None" and message_id != "":
+                    try:
+                        message = await logging_channel.fetch_message(int(message_id))
+                        if message:
+                            await message.delete()
+                    except discord.errors.NotFound:
+                        self.log.debug(guild_id, f"{self._module}.{_method}", f"Message {message_id} not found in channel {logging_channel}")
+
+            # remove all tracked items for this live platform (should only be one)
+            self.db.untrack_live(guild_id, user.id, platform)
+            await self.remove_live_roles(user, cog_settings)
 
     def find_platform_emoji(self, guild: discord.Guild, platform: str) -> typing.Union[discord.Emoji, None]:
         if guild is None:
@@ -374,7 +397,6 @@ class LiveNow(commands.Cog):
         if twitch_user:
             result = requests.get(f"http://decapi.me/twitch/avatar/{twitch_user}")
             if result.status_code == 200:
-                self.log.debug(0, f"{self._module}.{_method}", f"Got profile image for {twitch_user}: {result.text}")
                 return result.text
             else:
                 self.log.debug(0, f"{self._module}.{_method}", f"Failed to get profile image for {twitch_user}")
